refactor(confluence): add user operations mixin and integrate user details methods
- Introduced a new `UsersMixin` class for handling Confluence user operations, including methods to fetch user details by account ID and retrieve current user information.
- Updated `ConfluenceFetcher` to include `UsersMixin`, enhancing its functionality.
- Removed redundant user detail methods from `ConfluenceClient` to streamline the codebase and improve maintainability.
- Adjusted unit tests to utilize the updated `ConfluenceFetcher` for user detail retrieval.
